tidy: Move generic query functions to own file
